+++ src/modules/shared/components/CertificateIssuanceController/CertificateIssuanceController.tsx
@@ -1,9 +1,6 @@
-import { useRef, useState } from 'react';
-
-import { QRCodeCanvas } from 'qrcode.react';
+import { useState } from 'react';
 
 import { useModalController } from '../../hooks/useModalController';
-import { getPublicTokenPageURL } from '../../utils/getPublicTokenPageURL';
 import { CertificateIssuanceModal } from '../CertificateIssuanceModal';
 import { ConfirmationCertificateIssuedModal } from '../ConfirmationCertificateIssuedModal';
 
@@ -19,23 +16,14 @@ interface Props {
   chainId: number;
   tokenId: string;
 }
-
 export const CertificateIssuanceController = ({
   isOpen,
   onClose,
   contractAddress,
-  description,
-  image,
-  name,
-  transactionHash,
-  originalOwnerWalletAddress,
   chainId,
   tokenId,
 }: Props) => {
   const [finished, setFinished] = useState(false);
-  const qrCodeRef = useRef<HTMLDivElement>(null);
-  const canvas = qrCodeRef.current?.children[0] as HTMLCanvasElement;
-  const qrCodeSrc = canvas?.toDataURL();
 
   const {
     isOpen: isOpenConfirmationModal,
@@ -56,25 +44,12 @@ export const CertificateIssuanceController = ({
 
   return (
     <>
-      <div ref={qrCodeRef} className="pw-absolute pw-left-0 -pw-top-[500px]">
-        <QRCodeCanvas
-          width={70}
-          height={70}
-          value={`${
-            process.env.NEXT_PUBLIC_APP_BASE_PATH
-          }${getPublicTokenPageURL({ chainId, contractAddress, tokenId })}`}
-        />
-      </div>
       {isOpen ? (
         <>
           <CertificateIssuanceModal
-            QRCodeSrc={qrCodeSrc ?? ''}
-            name={name}
-            description={description}
             contractAddress={contractAddress}
-            image={image}
-            transactionHash={transactionHash}
-            originalOwnerWalletAddress={originalOwnerWalletAddress}
+            tokenId={tokenId}
+            chainId={chainId}
             isOpen={isOpen && !finished}
             onClose={onClose}
             onConfirm={handleConfirm}